requirements.txt
streamlit
openai>=1.40.0
python-dotenv
requests
